refactoring commit
renamed functions folder to constants
