+++ lib/presentation/home/screens/tabs/home_tab/home_tab.dart
@@ -8,7 +8,7 @@ import 'package:dvt_weather/presentation/constants/constants.dart';
 import 'package:dvt_weather/res/colors/colors.dart';
 import 'package:dvt_weather/presentation/home/screens/tabs/home_tab/widgets/current_weather_tile.dart';
 import 'package:dvt_weather/presentation/home/screens/tabs/home_tab/widgets/forecast_list.dart';
-import 'package:dvt_weather/res/functions/functions.dart';
+import 'package:dvt_weather/res/constants/functions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
