+++ lib/presentation/home/screens/tabs/home_tab/widgets/current_weather_display.dart
@@ -1,9 +1,9 @@
 import 'package:dvt_weather/controller/bloc/current_weather/current_weather_bloc.dart';
+import 'package:dvt_weather/presentation/constants/constants.dart';
+import 'package:dvt_weather/res/constants/functions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../../../res/functions/functions.dart';
-import '../../../../../constants/constants.dart';
 
 class CurrentWeatherDisplay extends StatelessWidget {
   const CurrentWeatherDisplay({super.key});
