+++ lib/res/constants/api_key.dart
@@ -1,0 +1,4 @@
+
+// I was going to add this to git ignore but I figured you might
+// want to run the app on your own. :)
+final apiKey = "e9386b4dcfd4708d672b76dfd60f9f82";
